Show all revisions for all branches with more than one head
In order to clean up a source repository it is useful to
see all revisions on all branches before the program quits.

Also display "Changeset ID" instead of "Revision number"
since revision numbers can be different on different copies
of a repository. Revision numbers also tends to change when
"hg strip" is used.